#380: Refactor a little.
