+++ ladder/ladder-parser/src/main/scala/com/actionfps/ladder/parser/ByteBufferExtension.scala
@@ -1,0 +1,16 @@
+package com.actionfps.ladder.parser
+
+import java.nio.ByteBuffer
+
+import scala.annotation.tailrec
+
+/**
+  * Created by william on 14/5/17.
+  */
+case class ByteBufferExtension(byteBuffer: ByteBuffer) {
+
+
+}
+object ByteBufferExtension {
+
+}
